Seperate arguments + only run if results.csv not present

+++ fit_all.py
@@ -47,10 +47,13 @@ FIP_elements = ["Mg", "Si"]
 min_E = 1.3
 max_E = 10.0
 bin_size = "27S"
-cutoff_cps = 1.
+xsm_cutoff_cps = 1.
+daxss_cutoff_cps = 10.
 
 do_dynamic_elements = True
-fit_args = {"min_E": min_E, "max_E": max_E, "do_dynamic_elements": do_dynamic_elements,"cutoff_cps":cutoff_cps}
+xsm_fit_args = {"min_E": min_E, "max_E": max_E, "do_dynamic_elements": do_dynamic_elements,"cutoff_cps":xsm_cutoff_cps}
+daxss_fit_args = {"min_E": min_E, "max_E": max_E, "do_dynamic_elements": do_dynamic_elements,"cutoff_cps":daxss_cutoff_cps}
+model_args = {"FIP_elements": FIP_elements}
 #%% Instrument specifics
 # xsm Stuff
 xsm_folder = "./data/xsm/data/"
@@ -64,36 +67,38 @@ daxss.load_data(DAXSS_file, daxss_rmf_path, daxss_arf_path)
 for flare_num in flares:
     for instrument_i in instruments_list:
         flare_dir = f"./data/pha_class/flare_num_{flare_num}/{instrument_i.name}"
-        xsm.set_output_dir(flare_dir)
+        fit_file = f"{flare_dir}/fit/results.csv"
+        instrument_i.set_output_dir(flare_dir)
         flare = observation_table.loc[flare_num]
         time_beg = flare["event_starttime"]
         time_end = flare["event_endtime"]
         #%% Create files if it does not exist
-        if not os.path.isdir(f"{flare_dir}/orig_pha/"):
-            CREATE = True
-        # else:
-            # PHA_file_list = glob.glob(f"{flare_dir}/orig_pha/*.pha")
-            # instrument_i.set_pha_files(PHA_file_list, ["USE_DEFAULT"] * len(PHA_file_list))
-        if CREATE:
-            shutil.rmtree(f"{flare_dir}/orig_pha/", ignore_errors=True)
-            instrument_i.create_pha_files(time_beg, time_end, bin_size=bin_size)
+        if not os.path.isfile(fit_file):
+            if not os.path.isdir(f"{flare_dir}/orig_pha/"):
+                CREATE = True
+            # else:
+                # PHA_file_list = glob.glob(f"{flare_dir}/orig_pha/*.pha")
+                # instrument_i.set_pha_files(PHA_file_list, ["USE_DEFAULT"] * len(PHA_file_list))
+            if CREATE:
+                shutil.rmtree(f"{flare_dir}/orig_pha/", ignore_errors=True)
+                instrument_i.create_pha_files(time_beg, time_end, bin_size=bin_size)
 
-        if not os.path.isdir(f"{flare_dir}/grpd_pha/"):
-            BIN = True
-        if BIN:
-            shutil.rmtree(f"{flare_dir}/grpd_pha/", ignore_errors=True)
-            instrument_i.do_grouping(min_count)
-        else:
-            PHA_file_list = glob.glob(f"{flare_dir}/grpd_pha/*.pha")
+            if not os.path.isdir(f"{flare_dir}/grpd_pha/"):
+                BIN = True
+            if BIN:
+                shutil.rmtree(f"{flare_dir}/grpd_pha/", ignore_errors=True)
+                instrument_i.do_grouping(min_count)
+            else:
+                instrument_i.PHA_file_list = glob.glob(f"{flare_dir}/grpd_pha/*.pha")
+            #%% Fit
             if instrument_i.name == "daxss":
-                instrument_i.set_pha_files(PHA_file_list, ["USE_DEFAULT"] * len(PHA_file_list))
+                instrument_i.set_pha_files(instrument_i.PHA_file_list, ["USE_DEFAULT"] * len(instrument_i.PHA_file_list))
+                instrument_i.fit(chisoth_2T, model_args, daxss_fit_args)
+                plot_individual(instrument_i.name,flare_dir,)
             if instrument_i.name == "xsm":
-                instrument_i.set_pha_files(PHA_file_list, instrument_i.arf_files)
-
-        #%% Fit
-        # Create model arguments
-        model_args = {"FIP_elements": FIP_elements}
-        instrument_i.fit(chisoth_2T, model_args, fit_args)
-        plot_individual(instrument_i.name,flare_dir,)
-    plot_simult(flare_num,instrument_names)
+                instrument_i.set_pha_files(instrument_i.PHA_file_list, instrument_i.arf_files)
+                instrument_i.fit(chisoth_2T, model_args, xsm_fit_args)
+                plot_individual(instrument_i.name,flare_dir,)
+            # Create model arguments
+    plot_simult("./data/pha_class",flare_num,instrument_names)
